Move FsUtils to fsUtils.ts
