+++ src/targets/browser/browserPathResolver.ts
@@ -24,7 +24,7 @@ import { injectable, inject } from 'inversify';
 import { IVueFileMapper, VueHandling } from '../../adapter/vueFileMapper';
 import { ILogger } from '../../common/logging';
 import { IFsUtils } from '../../common/fsUtils';
-import { FSUtils } from '../../ioc-extras';
+import { FSUtils } from '../../common/fsUtils';
 
 interface IOptions extends ISourcePathResolverOptions {
   baseUrl?: string;
