+++ src/targets/node/nodeAttacher.ts
@@ -23,7 +23,7 @@ import { IProgram, StubProgram, WatchDogProgram } from './program';
 import { IRestartPolicy, RestartPolicyFactory } from './restartPolicy';
 import { WatchDog } from './watchdogSpawn';
 import { LocalFsUtils } from '../../common/fsUtils';
-import { FSUtils } from '../../ioc-extras';
+import { FSUtils } from '../../common/fsUtils';
 
 const localize = nls.loadMessageBundle();
